[#36]feature-#36:소모임 목록 조회 Paging 처리 추가

+++ src/main/java/com/flab/daily/mapper/MeetingMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Mapper
 public interface MeetingMapper {
     int addMeeting(MeetingDAO meetingDAO);
-    int countMeetingAll();
+    Long countMeetingAll();
     List<MeetingResponseDTO> findMeetingList(Pagination pagination);
     MeetingResponseDTO findMeetingOneById(Long meetingId);
 }
